+++ R/doFilter.R
@@ -346,6 +346,9 @@ filterPDTMT <- function(sce, inputLevel, minScore = 0, minPeptides = 0,
 #' @param plotUpset Logical scalar, whether to generate an UpSet plot
 #'     detailing the reasons for features being filtered out. Only
 #'     generated if any feature is in fact filtered out.
+#' @param revPattern Character scalar providing the pattern (a regular
+#'     expression) used to identify decoys (reverse hits). The pattern is
+#'     matched against the IDs in the FragPipe \code{Protein} column.
 #' @param exclFile Character scalar, the path to a text file where the
 #'     features that are filtered out are written. If \code{NULL} (default),
 #'     excluded features are not recorded.
@@ -358,15 +361,16 @@ filterPDTMT <- function(sce, inputLevel, minScore = 0, minPeptides = 0,
 #' @importFrom rlang .data
 #'
 filterFragPipe <- function(sce, minPeptides, plotUpset = TRUE,
-                           exclFile = NULL) {
+                           revPattern = "^rev_", exclFile = NULL) {
     .assertVector(x = sce, type = "SummarizedExperiment")
     .assertScalar(x = minPeptides, type = "numeric", allowNULL = TRUE)
     .assertScalar(x = plotUpset, type = "logical")
+    .assertScalar(x = revPattern, type = "character")
     .assertScalar(x = exclFile, type = "character", allowNULL = TRUE)
 
     ## Make sure that the columns used for filtering later are character vectors
     rowData(sce)$Potential.contaminant <- ifelse(grepl("^contam_", rowData(sce)$Protein), "+", "")
-    rowData(sce)$Reverse <- ifelse(grepl("^rev_", rowData(sce)$Protein), "+", "")
+    rowData(sce)$Reverse <- ifelse(grepl(revPattern, rowData(sce)$Protein), "+", "")
 
     filtdf <- as.data.frame(SummarizedExperiment::rowData(sce)) %>%
         dplyr::select(dplyr::any_of(c("Reverse", "Potential.contaminant",